refactor(etl): move osc dataset utilities to new package
